install_k8s/keycloak/requirements-cli.txt
python-dotenv
requests
python-jose